requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-cors==0.7.0
aiohttp_remotes==1.2.0
aiopg==1.4.0
aiosignal==1.2.0
alembic==1.14.0
annotated-types==0.7.0
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.4.1
attrs==24.3.0
chardet==5.0.0
charset-normalizer==2.1.0
colorama==0.4.5
exceptiongroup==1.2.2
execnet==2.1.1
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.2.1
Markdown==3.7
MarkupSafe==2.1.1
more-itertools==8.13.0
multidict==6.0.2
mypy==1.13.0
mypy-extensions==1.0.0
packaging==21.3
pluggy==1.5.0
propcache==0.2.1
psycopg2-binary==2.9.10
py==1.11.0
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==8.3.4
pytest-aiohttp==1.0.5
pytest-asyncio==0.19.0
pytest-forked==1.4.0
pytest-xdist==3.6.1
python-dateutil==2.8.2
python-editor==1.0.4
referencing==0.35.1
rpds-py==0.22.3
six==1.16.0
SQLAlchemy==1.4.54
toml==0.10.2
tomli==2.0.1
types-jsonschema==4.23.0.20241208
types-Markdown==3.7.0.20241204
types-psycopg2==2.9.21.20241019
types-SQLAlchemy==1.4.53.38
typing_extensions==4.12.2
wcwidth==0.2.5
yarl==1.18.3
zipp==3.8.1
